+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Changelog
 * FEATURE: link to internal section references
 * FEATURE: advanced work filtering (publication, commencement, repeal, amendment etc.)
 * FEATURE: show offline warning when editing a document
+* FEATURE: site sidebar removed and replaced with tabs
 * New schedule syntax makes headings and subheadings clearer
 * Move document templates from templates/documents/ to templates/indigo_api/documents/